update smart contract

+++ contracts/ZombieFactory.sol
@@ -5,7 +5,6 @@
 pragma solidity ^0.8.0;
 
 contract ZombieFactory {
-
     // Events are a way for the contract to communicate that something happened on the blockchain
     // to the front-end app, which can be 'listening' for certain events and take action when they
     // happen.
@@ -18,19 +17,19 @@ contract ZombieFactory {
     //   return result;
     // }
     // Then the front-end app can listen to the event like this:
-    // MyContract.IntegersAdded((error, result) => {...}); 
-    event NewZombie(uint zombieId, string name, uint dna);
+    // MyContract.IntegersAdded((error, result) => {...});
+    event NewZombie(uint256 zombieId, string name, uint256 dna);
 
     // State variable that will be stored permanently in the Ethereum blockchain.
     // The uint type corresponds to an unsigned integer, meaning its value must be non-negative.
     // Its size is 256 bits but it's possible to declare uints with less bits (uint8, unit16...).
-    uint dnaDigits = 16;
-    uint dnaModulus = 10 ** dnaDigits;
+    uint256 dnaDigits = 16;
+    uint256 dnaModulus = 10**dnaDigits;
 
     // Structs allow to create more complicated data types
     struct Zombie {
         string name;
-        uint dna;
+        uint256 dna;
     }
 
     // There are two types of arrays in Solidity: fixed length (uint[2] fixedArray) and dynamic
@@ -40,8 +39,8 @@ contract ZombieFactory {
 
     // Mappings are another way of storing organized data, it's a key-value store for storing and
     // looking up data
-    mapping(uint => address) public zombieToOwner;
-    mapping(address => uint) ownerZombieCount;
+    mapping(uint256 => address) public zombieToOwner;
+    mapping(address => uint256) ownerZombieCount;
 
     // 1) Functions, as variables, also have a visibility.
     // By default, functions are public meaning that anyone (or any other contract) can call the
@@ -69,10 +68,10 @@ contract ZombieFactory {
     // - view function if the function is only viewing the data and not modifying it.
     // - pure function if the function is not even viewing the data of the application.
     //   ie. function _add(uint a, uint b) private pure returns (uint) { return a + b; }
-    function _createZombie(string memory _name, uint _dna) internal {
+    function _createZombie(string memory _name, uint256 _dna) internal {
         // array.push() adds something at the end of the array
         zombies.push(Zombie(_name, _dna));
-        uint id = zombies.length - 1;
+        uint256 id = zombies.length - 1;
         // msg.sender is a global variable available to all functions and it refers to the address
         // of the person (or smart contract) who called the current function. Since every contract
         // sits on the blockchain, waiting until someone calls one of its functions, there will
@@ -90,16 +89,20 @@ contract ZombieFactory {
     // slight change in the input will cause a large change in the hash. It also expects a single
     // parameter of type bytes, that's why we have to "pack" any parameters with abi.encodePacked()
     // before calling the keccak256 function.
-    function _generateRandomDna(string memory _str) private view returns (uint) {
+    function _generateRandomDna(string memory _str)
+        private
+        view
+        returns (uint256)
+    {
         // To cast a value into another data type, simply used type(value)
-        uint rand = uint(keccak256(abi.encodePacked(_str)));
+        uint256 rand = uint256(keccak256(abi.encodePacked(_str)));
         return rand % dnaModulus;
     }
 
     function createRandomZombie(string memory _name) public {
         // require will throw an error and stop executing if some condition is not true
         require(ownerZombieCount[msg.sender] == 0);
-        uint randDna = _generateRandomDna(_name);
+        uint256 randDna = _generateRandomDna(_name);
         _createZombie(_name, randDna);
     }
 }
